fix: refactor to multiple files
